requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
numpy==1.26.2
scikit-learn==1.3.2
apache-flink==1.18.0
python-dotenv==1.0.0
faker==21.0.0
